portfolio_bff/requirements.txt
Django>=4.2,<5.0
kafka-python>=2.0.2
PyMySQL>=1.1.0,<2.0
cryptography>=42.0.0,<43.0
